+++ src/sys/fs/device.rs
@@ -15,6 +15,7 @@ use crate::sys::net::socket::tcp::TcpSocket;
 use crate::sys::net::socket::udp::UdpSocket;
 use crate::sys::rng::Random;
 use crate::sys::speaker::Speaker;
+use crate::sys::keyboard::KeyboardBuffer;
 use crate::sys::vga::{VgaFont, VgaMode, VgaPalette, VgaBuffer};
 use crate::sys::snd::SoundBuffer;
 use crate::sys::process::{ProcId, ProcDir, ProcEnv, ProcUser};
@@ -52,6 +53,7 @@ pub enum DeviceType {
     ProcDir     = 22,
     ProcEnv     = 23,
     ProcUser    = 24,
+    KbdBuffer   = 25,
 }
 
 impl TryFrom<&[u8]> for DeviceType {
@@ -84,6 +86,7 @@ impl TryFrom<&[u8]> for DeviceType {
             22 => Ok(DeviceType::ProcDir),
             23 => Ok(DeviceType::ProcEnv),
             24 => Ok(DeviceType::ProcUser),
+            25 => Ok(DeviceType::KbdBuffer),
              _ => Err(()),
         }
     }
@@ -115,6 +118,7 @@ impl DeviceType {
             DeviceType::ProcDir     => ProcDir::size(),
             DeviceType::ProcEnv     => ProcEnv::size(),
             DeviceType::ProcUser    => ProcUser::size(),
+            DeviceType::KbdBuffer   => KeyboardBuffer::size(),
             _                       => 1,
         };
         let mut res = vec![0; len];
@@ -150,6 +154,7 @@ pub enum Device {
     ProcDir(ProcDir),
     ProcEnv(ProcEnv),
     ProcUser(ProcUser),
+    KbdBuffer(KeyboardBuffer),
 }
 
 impl TryFrom<&[u8]> for Device {
@@ -181,6 +186,7 @@ impl TryFrom<&[u8]> for Device {
             DeviceType::ProcDir     => Ok(Device::ProcDir(ProcDir::new())),
             DeviceType::ProcEnv     => Ok(Device::ProcEnv(ProcEnv::new())),
             DeviceType::ProcUser    => Ok(Device::ProcUser(ProcUser::new())),
+            DeviceType::KbdBuffer   => Ok(Device::KbdBuffer(KeyboardBuffer::new())),
             DeviceType::Drive if buf.len() > 2 => {
                 let bus = buf[1];
                 let dsk = buf[2];
@@ -255,6 +261,7 @@ impl FileIO for Device {
             Device::ProcDir(io)     => io.read(buf),
             Device::ProcEnv(io)     => io.read(buf),
             Device::ProcUser(io)    => io.read(buf),
+            Device::KbdBuffer(io)   => io.read(buf),
         }
     }
 
@@ -285,6 +292,7 @@ impl FileIO for Device {
             Device::ProcDir(io)     => io.write(buf),
             Device::ProcEnv(io)     => io.write(buf),
             Device::ProcUser(io)    => io.write(buf),
+            Device::KbdBuffer(io)   => io.write(buf),
         }
     }
 
@@ -315,6 +323,7 @@ impl FileIO for Device {
             Device::ProcDir(io)     => io.close(),
             Device::ProcEnv(io)     => io.close(),
             Device::ProcUser(io)    => io.close(),
+            Device::KbdBuffer(io)   => io.close(),
         }
     }
 
@@ -345,6 +354,7 @@ impl FileIO for Device {
             Device::ProcDir(io)     => io.poll(event),
             Device::ProcEnv(io)     => io.poll(event),
             Device::ProcUser(io)    => io.poll(event),
+            Device::KbdBuffer(io)   => io.poll(event),
         }
     }
 }