+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Add keyboard buffer device (#864)
 
 ## 0.13.0 (2026-06-21)
 - Fix memory leak during process creation ([#825](https://github.com/vinc/moros/pull/825))